requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
gunicorn==21.2.0
mongoengine==0.27.0
mssql-django==1.3
packaging==23.2
pymongo==3.12.0
pyodbc==5.0.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata @ file:///croot/python-tzdata_1690578112552/work
